Use interfaces for user services
